Create a shared HttpFixtures trait
