+++ archive/progress_http/src/test/scala/uk/ac/wellcome/platform/archive/progress_http/fixtures/ProgressHttpFixture.scala
@@ -2,8 +2,7 @@ package uk.ac.wellcome.platform.archive.progress_http.fixtures
 
 import java.net.URL
 
-import akka.http.scaladsl.Http
-import akka.http.scaladsl.model.{HttpEntity, HttpRequest, HttpResponse}
+import akka.http.scaladsl.model.HttpEntity
 import akka.stream.Materializer
 import io.circe.Decoder
 import org.scalatest.concurrent.ScalaFutures
@@ -11,7 +10,7 @@ import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.messaging.test.fixtures.SNS.Topic
 import uk.ac.wellcome.messaging.test.fixtures.{Messaging, SNS}
 import uk.ac.wellcome.platform.archive.common.config.models.HTTPServerConfig
-import uk.ac.wellcome.platform.archive.common.fixtures.RandomThings
+import uk.ac.wellcome.platform.archive.common.fixtures.{HttpFixtures, RandomThings}
 import uk.ac.wellcome.platform.archive.common.progress.fixtures.{
   ProgressGenerators,
   ProgressTrackerFixture
@@ -32,6 +31,7 @@ trait ProgressHttpFixture
     with ProgressTrackerFixture
     with ProgressGenerators
     with SNS
+    with HttpFixtures
     with Messaging {
 
   def withApp[R](table: Table,
@@ -96,12 +96,4 @@ trait ProgressHttpFixture
       .get
     fromJson[T](stringBody).get
   }
-
-  def whenRequestReady[R](r: HttpRequest)(testWith: TestWith[HttpResponse, R]): R =
-    withActorSystem { implicit actorSystem =>
-      val request = Http().singleRequest(r)
-      whenReady(request) { (response: HttpResponse) =>
-        testWith(response)
-      }
-    }
 }
